pyproject.toml
[tool.poetry]
name = "bsort"
version = "0.1.0"
description = "Bottle cap sorting ML CLI"
authors = ["Agatha <silalahiagatha15@gmail.com>"]
packages = [
  { include = "bsort", from = "src" }
]


[tool.poetry.dependencies]
python = ">=3.11,<3.12"
numpy = "^1.26"
opencv-python = "^4.8.0"
pyyaml = "^6.0"
wandb = "^0.15"

[tool.poetry.group.dev.dependencies]
black = "^24.9"
isort = "^5.12"
pylint = "^2.15"
pytest = "^7.4"
pre-commit = "^3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
